pyproject.toml
[project]
name = "ansible-collection-containers"
version = "1.5.0"
authors = [{ name = "VEXXHOST, Inc.", email = "support@vexxhost.com" }]
description = "Ansible collection for deploying container components"
readme = "README.md"
requires-python = ">=3.11"
license = "Apache-2.0"
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Environment :: OpenStack",
  "Intended Audience :: Information Technology",
  "Intended Audience :: System Administrators",
  "Operating System :: POSIX :: Linux",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
]
dependencies = [
  "ansible-core>=2.18",
  "docker-image-py>=0.1.12",
  "jmespath>=1.0.1",
]

[dependency-groups]
dev = [
  "molecule>=25.9.0.rc1",
  "pytest>=8.4.1",
  "pytest-forked>=1.6.0",
  "pytest-xdist>=3.7.0",
]

[project.urls]
source = "https://github.com/vexxhost/ansible-collection-containers"
issues = "https://github.com/vexxhost/ansible-collection-containers/issues"

[tool.uv]
package = false
